fix: do not ask PIMS to delete UF if it is failed (as it has been already done by PIMS)

+++ src/main/java/be/cytomine/service/image/UploadedFileService.java
@@ -403,7 +403,7 @@ public class UploadedFileService extends ModelService {
     protected void afterDelete(CytomineDomain domain, CommandResponse response) {
 
         UploadedFile uf = (UploadedFile) domain;
-        if (uf.getParent() == null) {
+        if (uf.getParent() == null && uf.getStatus() % 2 == 0) {
             try {
                 imageServerService.deleteFile(uf);
             } catch (Exception e) {
